+++ mapie/conformity_scores/residual_conformity_scores.py
@@ -24,6 +24,7 @@ class AbsoluteConformityScore(ConformityScore):
 
     def get_signed_conformity_scores(
         self,
+        X: ArrayLike,
         y: ArrayLike,
         y_pred: ArrayLike,
     ) -> NDArray:
@@ -36,14 +37,18 @@ class AbsoluteConformityScore(ConformityScore):
 
     def get_estimation_distribution(
         self,
+        X: ArrayLike,
         y_pred: ArrayLike,
-        conformity_scores: ArrayLike,
+        conformity_scores: ArrayLike
     ) -> NDArray:
         """
         Compute samples of the estimation distribution from the predicted
         values and the conformity scores, from the following formula:
         signed conformity score = y - y_pred
         <=> y = y_pred + signed conformity score
+
+        ``conformity_scores`` can be either the conformity scores or
+        the quantile of the conformity scores.
         """
         return np.add(y_pred, conformity_scores)
 
@@ -89,22 +94,21 @@ class GammaConformityScore(ConformityScore):
                 "in conformity with the Gamma distribution support."
             )
 
+    @staticmethod
     def _all_strictly_positive(
-        self,
         y: ArrayLike,
     ) -> bool:
-        if np.any(np.less_equal(y, 0)):
-            return False
-        return True
+        return not np.any(np.less_equal(y, 0))
 
     def get_signed_conformity_scores(
         self,
+        X: ArrayLike,
         y: ArrayLike,
         y_pred: ArrayLike,
     ) -> NDArray:
         """
-        Compute samples of the estimation distribution from the predicted
-        values and the conformity scores, from the following formula:
+        Compute the signed conformity scores from the observed values
+        and the predicted ones, from the following formula:
         signed conformity score = (y - y_pred) / y_pred
         """
         self._check_observed_data(y)
@@ -113,14 +117,18 @@ class GammaConformityScore(ConformityScore):
 
     def get_estimation_distribution(
         self,
+        X: ArrayLike,
         y_pred: ArrayLike,
-        conformity_scores: ArrayLike,
+        conformity_scores: ArrayLike
     ) -> NDArray:
         """
         Compute samples of the estimation distribution from the predicted
         values and the conformity scores, from the following formula:
         signed conformity score = (y - y_pred) / y_pred
         <=> y = y_pred * (1 + signed conformity score)
+
+        ``conformity_scores`` can be either the conformity scores or
+        the quantile of the conformity scores.
         """
         self._check_predicted_data(y_pred)
         return np.multiply(y_pred, np.add(1, conformity_scores))