+++ mapie/tests/test_conformity_scores.py
@@ -10,6 +10,7 @@ y_toy = np.array([5, 7, 9, 11, 13, 15])
 y_pred_list = [4, 7, 10, 12, 13, 12]
 conf_scores_list = [1, 0, -1, -1, 0, 3]
 conf_scores_gamma_list = [1 / 4, 0, -1 / 10, -1 / 12, 0, 3 / 12]
+random_state = 42
 
 
 class DummyConformityScore(ConformityScore):
@@ -17,12 +18,12 @@ class DummyConformityScore(ConformityScore):
         super().__init__(sym=True, consistency_check=True)
 
     def get_signed_conformity_scores(
-        self, y: ArrayLike, y_pred: ArrayLike,
+        self, X: ArrayLike, y: ArrayLike, y_pred: ArrayLike,
     ) -> NDArray:
         return np.subtract(y, y_pred)
 
     def get_estimation_distribution(
-        self, y_pred: ArrayLike, conformity_scores: ArrayLike
+        self, X: ArrayLike, y_pred: ArrayLike, conformity_scores: ArrayLike
     ) -> NDArray:
         """
         A positive constant is added to the sum between predictions and
@@ -45,10 +46,12 @@ def test_absolute_conformity_score_get_conformity_scores(
     """Test conformity score computation for AbsoluteConformityScore."""
     abs_conf_score = AbsoluteConformityScore()
     signed_conf_scores = abs_conf_score.get_signed_conformity_scores(
-        y_toy, y_pred
+        X_toy, y_toy, y_pred
     )
-    conf_scores = abs_conf_score.get_conformity_scores(y_toy, y_pred)
-    expected_signed_conf_scores = np.array([1, 0, -1, -1, 0, 3])
+    conf_scores = abs_conf_score.get_conformity_scores(
+        X_toy, y_toy, y_pred
+    )
+    expected_signed_conf_scores = np.array(conf_scores_list)
     expected_conf_scores = np.abs(expected_signed_conf_scores)
     np.testing.assert_allclose(signed_conf_scores, expected_signed_conf_scores)
     np.testing.assert_allclose(conf_scores, expected_conf_scores)
@@ -63,7 +66,9 @@ def test_absolute_conformity_score_get_estimation_distribution(
 ) -> None:
     """Test conformity observed value computation for AbsoluteConformityScore."""  # noqa: E501
     abs_conf_score = AbsoluteConformityScore()
-    y_obs = abs_conf_score.get_estimation_distribution(y_pred, conf_scores)
+    y_obs = abs_conf_score.get_estimation_distribution(
+        X_toy, y_pred, conf_scores
+    )
     np.testing.assert_allclose(y_obs, y_toy)
 
 
@@ -72,10 +77,10 @@ def test_absolute_conformity_score_consistency(y_pred: NDArray) -> None:
     """Test methods consistency for AbsoluteConformityScore."""
     abs_conf_score = AbsoluteConformityScore()
     signed_conf_scores = abs_conf_score.get_signed_conformity_scores(
-        y_toy, y_pred
+        X_toy, y_toy, y_pred
     )
     y_obs = abs_conf_score.get_estimation_distribution(
-        y_pred, signed_conf_scores
+        X_toy, y_pred, signed_conf_scores
     )
     np.testing.assert_allclose(y_obs, y_toy)
 
@@ -86,7 +91,9 @@ def test_gamma_conformity_score_get_conformity_scores(
 ) -> None:
     """Test conformity score computation for GammaConformityScore."""
     gamma_conf_score = GammaConformityScore()
-    conf_scores = gamma_conf_score.get_conformity_scores(y_toy, y_pred)
+    conf_scores = gamma_conf_score.get_conformity_scores(
+        X_toy, y_toy, y_pred
+    )
     expected_signed_conf_scores = np.array(conf_scores_gamma_list)
     np.testing.assert_allclose(conf_scores, expected_signed_conf_scores)
 
@@ -104,7 +111,9 @@ def test_gamma_conformity_score_get_estimation_distribution(
 ) -> None:
     """Test conformity observed value computation for GammaConformityScore."""  # noqa: E501
     gamma_conf_score = GammaConformityScore()
-    y_obs = gamma_conf_score.get_estimation_distribution(y_pred, conf_scores)
+    y_obs = gamma_conf_score.get_estimation_distribution(
+        X_toy, y_pred, conf_scores
+    )
     np.testing.assert_allclose(y_obs, y_toy)
 
 
@@ -113,10 +122,10 @@ def test_gamma_conformity_score_consistency(y_pred: NDArray) -> None:
     """Test methods consistency for GammaConformityScore."""
     gamma_conf_score = GammaConformityScore()
     signed_conf_scores = gamma_conf_score.get_signed_conformity_scores(
-        y_toy, y_pred
+        X_toy, y_toy, y_pred
     )
     y_obs = gamma_conf_score.get_estimation_distribution(
-        y_pred, signed_conf_scores
+        X_toy, y_pred, signed_conf_scores
     )
     np.testing.assert_allclose(y_obs, y_toy)
 
@@ -136,7 +145,9 @@ def test_gamma_conformity_score_check_oberved_value(
     """Test methods consistency for GammaConformityScore."""
     gamma_conf_score = GammaConformityScore()
     with pytest.raises(ValueError):
-        gamma_conf_score.get_signed_conformity_scores(y_toy, y_pred)
+        gamma_conf_score.get_signed_conformity_scores(
+            [], y_toy, y_pred
+        )
 
 
 @pytest.mark.parametrize(
@@ -148,6 +159,14 @@ def test_gamma_conformity_score_check_oberved_value(
     ],
 )
 @pytest.mark.parametrize(
+    "X_toy",
+    [
+        np.array([0, -7, 10, 12, 0, 12]).reshape(-1, 1),
+        np.array([0, 7, -10, 12, 1, -12]).reshape(-1, 1),
+        np.array([12, -7, 0, 12, 13, 2]).reshape(-1, 1),
+    ],
+)
+@pytest.mark.parametrize(
     "conf_scores",
     [
         np.array(conf_scores_gamma_list),
@@ -155,7 +174,7 @@ def test_gamma_conformity_score_check_oberved_value(
     ],
 )
 def test_gamma_conformity_score_check_predicted_value(
-    y_pred: NDArray, conf_scores: NDArray
+    y_pred: NDArray, conf_scores: NDArray, X_toy: NDArray
 ) -> None:
     """Test methods consistency for GammaConformityScore."""
     gamma_conf_score = GammaConformityScore()
@@ -163,27 +182,31 @@ def test_gamma_conformity_score_check_predicted_value(
         ValueError,
         match=r".*At least one of the predicted target is negative.*"
     ):
-        gamma_conf_score.get_signed_conformity_scores(y_toy, y_pred)
+        gamma_conf_score.get_signed_conformity_scores(
+            X_toy, y_toy, y_pred
+        )
     with pytest.raises(
         ValueError,
         match=r".*At least one of the predicted target is negative.*"
     ):
-        gamma_conf_score.get_estimation_distribution(y_pred, conf_scores)
+        gamma_conf_score.get_estimation_distribution(
+            X_toy, y_pred, conf_scores
+        )
 
 
 def test_check_consistency() -> None:
     """
-    Test that a dummy ConformityScore class that gives inconsistent conformity
-    scores and distributions raises an error.
+    Test that a dummy ConformityScore class that gives inconsistent scores
+    and distributions raises an error.
     """
     dummy_conf_score = DummyConformityScore()
     conformity_scores = dummy_conf_score.get_signed_conformity_scores(
-        y_toy, y_pred_list
+        X_toy, y_toy, y_pred_list
     )
     with pytest.raises(
         ValueError,
         match=r".*The two functions get_conformity_scores.*"
     ):
         dummy_conf_score.check_consistency(
-            y_toy, y_pred_list, conformity_scores
+            X_toy, y_toy, y_pred_list, conformity_scores
         )